chore: refactor random
